pyproject.toml
[project]
name = "youtube-api-mcp-server"
version = "1.0.0"
description = "Production-ready YouTube MCP Server using Google Data API v3 - 100% reliable alternative to yt-dlp"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Du'An Lightfoot", email = "noreply@labeveryday.com" }
]
keywords = ["youtube", "mcp", "api", "claude", "desktop", "extraction", "reliable"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Internet :: WWW/HTTP :: Dynamic Content",
    "Topic :: Multimedia :: Video",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]

dependencies = [
    "mcp>=1.0.0",
    "google-api-python-client>=2.100.0",
    "google-auth>=2.23.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "tenacity>=8.2.0",
    "yt-dlp>=2024.12.13",
    "cachetools>=5.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "mypy>=1.0.0",
    "pre-commit>=3.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/youtube_api_server"]

[tool.black]
line-length = 88
target-version = ["py311"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[project.scripts]
youtube-api-mcp-server = "youtube_api_server.server:main"

[project.urls]
Homepage = "https://github.com/labeveryday/youtube-api-project"
Documentation = "https://github.com/labeveryday/youtube-api-project#readme"
Repository = "https://github.com/labeveryday/youtube-api-project.git"
Issues = "https://github.com/labeveryday/youtube-api-project/issues"